testing: don't overwrite ErrataConnector._url
Currently testing fails for me being unable to find fixtures.  It
turns out that the `--stage` test overwrites ErrataConnector._url,
which then affects the rest of the tests.  I think this must have been
working before thanks to ordering luck, which probably put the test
above, which I think will reset this to the "production" site.
monkeypatch this so the value isn't overwritten.